backend/pyproject.toml
[tool.poetry]
name = "personal-assistant-backend"
version = "0.1.0"
description = "Django API"
authors = ["SerhiiMoskalenko"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.2"
djangorestframework = "^3.16"
django-environ = "^0.12.0"
pydantic-settings = "^2.9.1"
psycopg = { version = "^3.2", extras = ["binary"] }
alembic = "^1.15.2"
django-filter = "^25.1"
boto3 = "^1.38.13"
django-storages = {extras = ["boto3"], version = "^1.14.6"}
gunicorn = "^23.0.0"
celery = "^5.5.2"
redis = "^6.1.0"
django-cors-headers = "^4.7.0"
feedparser = "^6.0.11"
django-redis = "^5.4.0"


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
